added initial code for jwt authentication
