+++ Servers/src/types/IUserAuthRequest.ts
@@ -1,0 +1,5 @@
+import { Request } from "express"
+import { User } from "../models/user.model"
+export interface IUserAuthRequest extends Request {
+  user: User
+}
